refactor pseudocode

+++ db/migrate/20130902213431_create_etrade_table.rb
@@ -3,8 +3,8 @@ class CreateEtradeTable < ActiveRecord::Migration
     create_table :etrades do |t|
       t.timestamps
       t.datetime :date
-      t.float :balance, null: false
-      t.float :fees
+      t.decimal :balance, null: false
+      t.decimal :fees
     end
   end
 end
