+++ db/migrate/20130926032456_add_givshares_table.rb
@@ -2,13 +2,13 @@ class AddGivsharesTable < ActiveRecord::Migration
     def change
     create_table :givshares do |t|
       t.integer :charity_group_id, null: false
-      t.float :stripe_balance
-      t.float :etrade_balance
-      t.float :shares_outstanding_beginning
-      t.float :shares_bought_through_donations
-      t.float :shares_outstanding_end
-      t.float :donation_price
-      t.float :round_down_price
+      t.decimal :stripe_balance
+      t.decimal :etrade_balance
+      t.decimal :shares_outstanding_beginning
+      t.decimal :shares_bought_through_donations
+      t.decimal :shares_outstanding_end
+      t.decimal :donation_price
+      t.decimal :round_down_price
       t.timestamps
     end
