+++ db/migrate/20131001060903_add_share_total_and_grant_to_givshares.rb
@@ -1,11 +1,11 @@
 class AddShareTotalAndGrantToGivshares < ActiveRecord::Migration
   def change
-    add_column :givshares, :share_total, :float
-    add_column :givshares, :share_granted, :float
-    add_column :givshares, :donor_grant, :float
+    add_column :givshares, :share_total, :decimal
+    add_column :givshares, :share_granted, :decimal
+    add_column :givshares, :donor_grant, :decimal
     add_column :givshares, :is_grant, :integer
-    add_column :givshares, :etrade_adjustment, :float
-    add_column :givshares, :charity_group_balance, :float
+    add_column :givshares, :etrade_adjustment, :decimal
+    add_column :givshares, :charity_group_balance, :decimal
     add_column :givshares, :count, :integer
   end
 end
