+++ db/migrate/20131003040135_create_table_for_shares.rb
@@ -1,13 +1,13 @@
 class CreateTableForShares < ActiveRecord::Migration
   def change
     create_table :shares do |t|
-      t.float :stripe_balance
-      t.float :etrade_balance
-      t.float :share_total_beginning
-      t.float :shares_added_by_donation
-      t.float :shares_subtracted_by_grants
-      t.float :share_total_end
-      t.float :share_price
+      t.decimal :stripe_balance
+      t.decimal :etrade_balance
+      t.decimal :share_total_beginning
+      t.decimal :shares_added_by_donation
+      t.decimal :shares_subtracted_by_grants
+      t.decimal :share_total_end
+      t.decimal :share_price
       t.timestamps
     end
 
@@ -16,12 +16,12 @@ class CreateTableForShares < ActiveRecord::Migration
       t.integer :charity_group_id
       t.integer :dwolla_transaction_id
       t.date :date
-      t.float :shares_subtracted
+      t.decimal :shares_subtracted
       t.timestamps
     end
 
     add_column :donations, :shares_purchased, :string
-    add_column :donations, :shares_added, :float
+    add_column :donations, :shares_added, :decimal
     add_column :donations, :donor_id, :integer
 
   end
